requirements.txt
addfips==0.3.1
albumentations==1.3.0
asttokens==2.1.0
automium==0.2.6
automium-web==0.1.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
boto3==1.26.6
botocore==1.29.6
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
coloredlogs==15.0.1
commonmark==0.9.1
contourpy==1.0.6
coremltools==6.0
cycler==0.11.0
decorator==5.1.1
docopt==0.6.2
executing==1.2.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
flatbuffers==22.10.26
fonttools==4.38.0
humanfriendly==10.0
idna==3.4
imageio==2.22.4
importlib-metadata==4.12.0
importlib-resources==5.10.0
ipython==8.6.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
legacy==0.1.6
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mpmath==1.2.1
networkx==2.8.8
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
onnx==1.12.0
onnxruntime==1.13.1
onnxsim==0.4.9
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
openvino==2022.2.0
packaging==21.3
pafy==0.5.5
pandas==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pipreqs==0.4.11
prompt-toolkit==3.0.32
protobuf==4.21.2
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.3
pycocotools==2.0.6
Pygments==2.13.0
pyparsing==3.0.9
pyspark==3.2.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.28.1
rich==12.6.0
s3transfer==0.6.0
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.6.0
sympy==1.11.1
tflite-runtime==2.10.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2022.10.10
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.12
utils==1.0.1
wcwidth==0.2.5
Werkzeug==2.2.2
yarg==0.1.9
zipp==3.8.1
lxml
